==> 4. catchment-elect.ipynb <==
# Obtain Categorical Variables for Catchment Areas and Electoral Wards (R)

Output of Catchments and Wards csvs to be merged in merge.ipynb

# Import libraries
library(sf)
library(data.table)
library(ggplot2)
library(tmap)
library(tidygeocoder)
library(leaflet)
library(dplyr)
library(osmdata)
library(tmaptools)
library(tidyr)
library(readr)

## Load priceper postcodes
priceper = read_csv("schools_geo/Brighton_and_Hove_link_13082022.csv")
# Obtain unique postcodes
unique_postcodes <- unique(priceper$postcode)
postcode_dir <- data.frame(postcode = unique_postcodes)

# Check
nrow(postcode_dir)

## left join postcode_df to unique_postcodes 
brighton_postcodes <- left_join(postcode_dir, postcode_df, by=c("postcode" = "pcds"))
nrow(brighton_postcodes)
brighton_houses <- brighton_postcodes%>%
  #remove NAs or 0
  filter(long != 0 )%>%
  filter(lat != 0)%>%
  st_as_sf(., coords = c("long","lat"), 
                   crs = 4326)

brighton_houses <- st_transform(brighton_houses, crs = 27700)

## Read in geojson file of catchment areas
catchment_areas <- "schools_geo/shapefiles/BrightonSecondaryCatchments.geojson"
catchment_areas <- st_read(catchment_areas)
catchment_areas_sf <- st_as_sf(catchment_areas)

## Read in geojson file of electoral wards
wards <- "schools_geo/shapefiles/Wards_December_2022_Boundaries_GB_BFC_1176271257858185555.geojson"
wards <- st_read(wards)
wards_sf <- st_as_sf(wards)

# Perform a spatial join based on the spatial relationships (i.e., matching postcodes)
joined_sf <- st_join(brighton_houses, catchment_areas_sf, join = st_within)
head(joined_sf)

# Perform a spatial join based on the spatial relationships (i.e., matching postcodes)
joined_sf2 <- st_join(brighton_houses, wards_sf, join = st_within)
head(joined_sf2)

### Export
# Drop geometry
attributes_only <- st_drop_geometry(joined_sf)
# Drop id column
attributes_only$id <- NULL

# Export the data as a CSV file
write.csv(attributes_only, file = "data/BrightonSecondaryCatchments.csv", row.names = FALSE)

### Export
# Drop geometry
attributes_only2 <- st_drop_geometry(joined_sf2)
# Drop id column
attributes_only2$id <- NULL

# Export the data as a CSV file
write.csv(attributes_only2, file = "data/BrightonElectoralWards.csv", row.names = FALSE)

nrow(attributes_only2)
