==> 5. merge.ipynb <==
# Merging priceper (incl geo_services) with postcodes from schools (schools_geo/all_schools_geo_both.csv)


import pandas as pd
from sklearn.linear_model import LinearRegression
import statsmodels.api as sm
import statsmodels.formula.api as smf
import statsmodels
#import janitor

import numpy as np

import sklearn
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix

import matplotlib.pyplot as plt
import seaborn as sns
from statsmodels.graphics.gofplots import ProbPlot
import re

import warnings
warnings.filterwarnings("ignore")

## Merge priceper_addedgeo and schools_geo

priceper_geoservices = pd.read_csv("priceper_housing_data/added_geo (rail_cc)/priceper_railcc.csv")

all_schools_geo_both = pd.read_csv("schools/schools_geo/allschools_geo_both_new.csv")

#all_schools_geo_both = all_schools_geo_both.rename(columns={'id': 'postcode'})

list(priceper_geoservices)

list(all_schools_geo_both)

checktypes = all_schools_geo_both.dtypes
print(checktypes)

complete_data = priceper_geoservices.merge(all_schools_geo_both[['postcode','walk_nearest_sec', 'nearest_sec_name', 'walk_nearest_pri', 'nearest_pri_name']], on='postcode', how='left')

checktypes = complete_data.dtypes
print(checktypes)

complete_data['year'] = complete_data['year'].astype(str)

list(complete_data)

final_data = complete_data.copy()

def merge_priceper_years(final_data, all_schools_geo_both):
    for year in range(2000, 2020):            
        # Filter columns that start with the year prefix
        #year_cols = [col for col in all_schools_geo_both.columns if col.startswith(str(year))]
        year_pattern = re.compile(f"^{year}")
        year_cols = [col for col in all_schools_geo_both.columns if year_pattern.match(col)]
        for year_col in year_cols:
            # Merge the datasets based on postcode and year
            final_data = final_data.merge(all_schools_geo_both[['postcode', year_col]], 
                                            left_on=['postcode'], 
                                            right_on=['postcode'], 
                                            how='left')
    
    return final_data

final_data['year'] = final_data['year'].astype(str)
finalised_data = merge_priceper_years(final_data, all_schools_geo_both)

finalised_data.to_csv('final_data.csv', index=False)

## Merge IMD data

complete_data = pd.read_csv('final_data.csv')

imd2004 = pd.read_csv('imd/imd2004eng.csv')

imd2004.head()

geodata = pd.read_csv('geo/ONSPD_MAY_2023_UK/Data/multi_csv/ONSPD_MAY_2023_UK_BN.csv')
geodata = geodata[['pcds','lsoa01','lsoa11','lsoa21']]

complete_geo = complete_data.merge(geodata[['pcds','lsoa01','lsoa11','lsoa21']], 
                                            left_on=['postcode'], 
                                            right_on=['pcds'], 
                                            how='left')

#left merge left on lad21cd, right on 
complete_04 = complete_geo.merge(imd2004[['SOA','IMD SCORE']], 
                                            left_on=['lsoa01'], 
                                            right_on=['SOA'], 
                                            how='left')

# clean up
complete_04.drop('SOA', axis=1, inplace=True)
complete_04.rename(columns={'IMD SCORE': 'imd_2004'}, inplace=True)

imd2007 = pd.read_csv('imd/imd2007eng.csv')

#left merge left on lad21cd, right on 
complete_07 = complete_04.merge(imd2007[['LSOA','IMD SCORE']], 
                                            left_on=['lsoa01'], 
                                            right_on=['LSOA'], 
                                            how='left')

# clean up
complete_07.drop('LSOA', axis=1, inplace=True)
complete_07.rename(columns={'IMD SCORE': 'imd_2007'}, inplace=True)

imd2010 = pd.read_csv('imd/imd2010eng.csv')

#left merge left on lad21cd, right on 
complete_10 = complete_07.merge(imd2010[['lsoacode','imd_score']], 
                                            left_on=['lsoa01'], 
                                            right_on=['lsoacode'], 
                                            how='left')

# clean up
complete_10.drop('lsoacode', axis=1, inplace=True)
complete_10.rename(columns={'imd_score': 'imd_2010'}, inplace=True)

imd2015 = pd.read_csv('imd/imd2015eng.csv')

#left merge left on lad21cd, right on 
complete_15 = complete_10.merge(imd2015[['LSOA code (2011)','Index of Multiple Deprivation (IMD) Score']], 
                                            left_on=['lsoa01'], 
                                            right_on=['LSOA code (2011)'], 
                                            how='left')

# clean up
complete_15.drop('LSOA code (2011)', axis=1, inplace=True)
complete_15.rename(columns={'Index of Multiple Deprivation (IMD) Score': 'imd_2015'}, inplace=True)

#CDRC Harmonised IMD 2019
imd2019 = pd.read_csv('imd/uk_imd2019.csv')
##note its rank and SOA_decile

#left merge left on lad21cd, right on 
complete_19 = complete_15.merge(imd2019[['LSOA','Rank','SOA_decile']], 
                                            left_on=['lsoa11'], 
                                            right_on=['LSOA'], 
                                            how='left')

# clean up
complete_19.drop('LSOA', axis=1, inplace=True)
complete_19.rename(columns={'Rank': 'imd_2019_rank'}, inplace=True)
complete_19.rename(columns={'SOA_decile': 'imd_2019_dec'}, inplace=True)

####complete_19.to_csv('final_imd_data.csv')

## Add Catchment Area variable (sec schools)

main = pd.read_csv('final_imd_data.csv')

catchment = pd.read_csv('schools/BrightonSecondaryCatchments.csv')

main.head()

catchment.head()

merged = main.merge(catchment[['postcode','AreaName']], on='postcode', how='left')
merged.head(20)

merged.rename(columns={'AreaName': 'SecCatchment'}, inplace=True)

merged.to_csv('final_imd_catch.csv', index=False)

## Add Electoral Ward


main = pd.read_csv('2000_cleaned_data_final_oac11.csv')

wards = pd.read_csv('schools/BrightonElectoralWards.csv')

main.head()

wards.head()

print(wards["WD22NM"].unique())

print(wards["WD22NM"].value_counts())

print(wards["WD22NM"].isna().sum())

merged = main.merge(wards[['postcode','WD22NM']], on='postcode', how='left')
merged.head(20)

print(merged["WD22NM"].value_counts())

merged.rename(columns={'WD22NM': 'ElectWards'}, inplace=True)

merged.to_csv('2000_cleaned_data_final_oac11_elec.csv', index=False)
